backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Resume Processing
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Web Scraping (for future phases)
playwright==1.40.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
